refactor: drop the leading underscore from internal module filenames
Rename _async_runtime.py -> async_runtime.py, _agent_limit.py ->
agent_limit.py, and _kernel_bootstrap.py -> kernel_bootstrap.py (via
git mv, so history follows), updating all imports and doc references.
These modules are internal to the rlm package; the private-file
convention added noise without guarding any external surface.
